% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hit_song.R
\name{hit_song}
\alias{hit_song}
\title{Hit song generator}
\usage{
hit_song(v = 4, c = 4, b = 4)
}
\arguments{
\item{v}{numeric. Number of lines in the verse.}

\item{c}{numeric. Number of lines in the chorus.}

\item{b}{numeric. Number of lines in the bridge.}
}
\description{
This function allows you to generate a guaranteed #1 hit with no effort!
}
