% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyrics.R
\docType{data}
\name{lyrics}
\alias{lyrics}
\title{Lyrics from Beyonce and Taylor Swift songs}
\format{
A data frame with 23084 rows and 2 variables:
\describe{
\item{artist}{name of the artist}
\item{line}{a line from a Beyonce or Taylor Swift song}
...
}
}
\source{
credit: Rosie Baillie and Dr. Sara Stoudt.
}
\usage{
lyrics
}
\description{
A dataset containing lyrics from Beyonce and Taylor Swift songs.
}
\keyword{datasets}
